+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Updated rhino3dm to be based on the Rhino 8 version of OpenNURBS
 - (py) Linux version now includes draco support
-- (py) Added Line.Transform
+- (py) Added Line.Transform, Brep.TryConvertBrep
 - (.net) Added macOS arm64 (Apple Silicon) builds
 
 ### Changed